requirements.txt
setuptools==65.5.0
wheel==0.38.4
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.4.1
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
EasyProcess==1.1
entrypoint2==1.1
Farama-Notifications==0.0.4
filelock==3.12.4
flatten-dict==0.4.2
fonttools==4.42.1
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pygame==2.5.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyriemann==0.5
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sb3-contrib==1.6.2
scikit-learn==1.3.1
scipy==1.10.1
sentry-sdk==1.31.0
setproctitle==1.3.2
Shimmy==1.2.1
six==1.16.0
smmap==5.0.0
stable-baselines3==1.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
TextGrid==1.5
threadpoolctl==3.2.0
tk==0.1.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.10
Werkzeug==2.3.7
zipp==3.16.2